group slow running tests in regression folder
